chore: change popular review location
